+++ client-common/src/test/java/org/realityforge/replicant/client/transport/TestDataLoadService.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.Set;
+import java.util.List;
 import java.util.function.Consumer;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -234,6 +235,29 @@ public final class TestDataLoadService
                                             @Nonnull final Object filterParameter,
                                             @Nonnull final Consumer<Runnable> completionAction,
                                             @Nonnull final Consumer<Runnable> failAction )
+  {
+  }
+
+  @Override
+  protected void requestBulkSubscribeToGraph( @Nonnull final List<ChannelDescriptor> descriptor,
+                                              @Nullable final Object filterParameter,
+                                              @Nonnull final Consumer<Runnable> completionAction,
+                                              @Nonnull final Consumer<Runnable> failAction )
+  {
+  }
+
+  @Override
+  protected void requestBulkUnsubscribeFromGraph( @Nonnull final List<ChannelDescriptor> descriptors,
+                                                  @Nonnull final Consumer<Runnable> completionAction,
+                                                  @Nonnull final Consumer<Runnable> failAction )
+  {
+  }
+
+  @Override
+  protected void requestBulkUpdateSubscription( @Nonnull final List<ChannelDescriptor> descriptors,
+                                                @Nonnull final Object filterParameter,
+                                                @Nonnull final Consumer<Runnable> completionAction,
+                                                @Nonnull final Consumer<Runnable> failAction )
   {
   }